pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "A utility for comparing configuration files"
authors = ["artxnv <artxnv@gmail.com>"]
readme = "README.md"
packages = [
    { include = "gendiff" }
]

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.dependencies]
python = "^3.11"
PyYAML = "^6.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
isort = "^5.13.2"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"